+++ fortanix-vme/fortanix-vme-runner/Cargo.toml
@@ -4,11 +4,9 @@ version = "0.1.0"
 edition = "2018"
 authors = ["Fortanix, Inc."]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 fortanix-vme-abi = { path = "../fortanix-vme-abi" }
 nix = "0.22.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_cbor = { version = "0.11" }
-vsock = "0.2.5"
+vsock = "0.2.4"
