+++ fortanix-vme/fortanix-vme-runner/src/main.rs
@@ -1,8 +1,7 @@
-use fortanix_vme_runner::server::Server;
+use fortanix_vme_runner::Server;
 use fortanix_vme_abi::SERVER_PORT;
 
 fn main() {
-    let (server_thread, _port) = Server::run(SERVER_PORT)
-                                        .expect("Server failed");
+    let (server_thread, _port) = Server::run(SERVER_PORT).expect("Server failed");
     server_thread.join().expect("Server panicked");
 }
